backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
websockets
requests
google-cloud-aiplatform
